fix: apply 50+ audit fixes across settlement, materials, VFX, security, interiors
- Settlement generator: room_functions, building floors, cluster maps, lot clipping
- Materials: cobblestone key fix, glass/water recipes, bronze patina, dead Musgrave
- VFX templates: sanitize_cs_identifier fixes, PrimeTween guards, deprecated casts
- Security: remove type from BLOCKED_FUNCTIONS per user policy
- Interior binding: add 9 Hearthvale building types to BUILDING_ROOM_MAP
- Tool runner: fix deprecated tempfile.mktemp, add logger
- Unity templates: fix IDamageable, projectile damage, monster avatar type
- Export: add GLTF PBR settings, animation timing sidecar
- 293 new tests added across 11 test files

+++ Tools/mcp-toolkit/src/veilbreakers_mcp/shared/unity_templates/monster_vfx_templates.py
@@ -750,7 +750,7 @@ public class MonsterSignatureVFXController : MonoBehaviour
         if (config.BurstCount > 0)
         {{
             emission.SetBursts(new ParticleSystem.Burst[] {{
-                new ParticleSystem.Burst(0f, (short)(config.BurstCount * intensityMultiplier))
+                new ParticleSystem.Burst(0f, (int)(config.BurstCount * intensityMultiplier))
             }});
         }}
 
@@ -1208,7 +1208,7 @@ public class HybridBrandVFXController : MonoBehaviour
         if (config.BurstCount > 0)
         {{
             emission.SetBursts(new ParticleSystem.Burst[] {{
-                new ParticleSystem.Burst(0f, (short)config.BurstCount)
+                new ParticleSystem.Burst(0f, (int)config.BurstCount)
             }});
         }}
 
@@ -1844,7 +1844,7 @@ public class ExecuteThresholdVFXController : MonoBehaviour
 
         var emission = brandExecutePS.emission;
         emission.SetBursts(new ParticleSystem.Burst[] {{
-            new ParticleSystem.Burst(0f, (short)config.BurstCount)
+            new ParticleSystem.Burst(0f, (int)config.BurstCount)
         }});
 
         // Inward pull for drain-type brands
